refactor: inlineCompletionHandler
